lang_channel/requirements.txt
# I want to use last versions of libraries
fastapi[all]
python-telegram-bot==20.0a2
Pillow
gspread

# dev
pytest
